+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- updateOrderStatus event handler
+
+### Changed
+
+- Change setAffiliatesOrders event topic to order-created
+
 ## [0.2.0] - 2022-01-03
 
 ### Added